basic image processing
This code loads the image from the specified file stream using imageio loading and then prints the dimensions (shape) of the loaded image to the screen. This returns a tuple containing the image's width, differences, and number of color details. For example, an output of (height, width, color_channels) will not be obtained. This information helps you understand the image's dimensions and number of color channels.